fix: eth: use the correct tipset for Eth APIs
Many of our "state" methods look at the tipset's _parent_ state.
However, Eth APIs expect the current tipset. We had sort-of fixed this
for balances, but nothing else.

This patch alters _most_ Eth APIs to use the following tipset as their
base instead of the requested tipset. This is now possible because the
pending tipset is no longer allowed.